fix(rpc/v05/call): return ContractErrorV05 for reverted execution
